+++ src/tests/test.index.js
@@ -1,38 +1,38 @@
-const Eth = require('../index.js'); // eslint-disable-line
+const Vap = require('../index.js'); // eslint-disable-line
 const assert = require('chai').assert; // eslint-disable-line
-const util = require('ethjs-util');
-const EthQuery = require('eth-query'); // eslint-disable-line
-const TestRPC = require('ethereumjs-testrpc');
+const util = require('vapjs-util');
+const VapQuery = require('vap-query'); // eslint-disable-line
+const TestRPC = require('vaporyjs-testrpc');
 const BigNumber = require('bignumber.js'); // eslint-disable-line
-const abi = require('ethjs-abi'); // eslint-disable-line
+const abi = require('vapjs-abi'); // eslint-disable-line
 const provider = TestRPC.provider({});
 
-describe('ethjs-query', () => {
+describe('vapjs-query', () => {
   describe('construction', () => {
     it('should construct normally', () => {
-      const eth = new Eth(provider);
+      const vap = new Vap(provider);
 
-      assert.equal(typeof eth, 'object');
-      assert.equal(typeof eth.query, 'object');
-      assert.equal(typeof eth.accounts, 'function');
-      assert.equal(typeof eth.getBalance, 'function');
-      assert.equal(typeof eth.sendTransaction, 'function');
-      assert.equal(typeof eth.sendRawTransaction, 'function');
+      assert.equal(typeof vap, 'object');
+      assert.equal(typeof vap.query, 'object');
+      assert.equal(typeof vap.accounts, 'function');
+      assert.equal(typeof vap.getBalance, 'function');
+      assert.equal(typeof vap.sendTransaction, 'function');
+      assert.equal(typeof vap.sendRawTransaction, 'function');
     });
 
-    it('should check if the rpc is eth_syncing', () => {
-      const eth = new Eth(provider);
+    it('should check if the rpc is vap_syncing', () => {
+      const vap = new Vap(provider);
 
-      eth.syncing((err, result) => {
+      vap.syncing((err, result) => {
         assert.equal(err, null);
         assert.equal(typeof result, 'boolean');
       });
     });
 
-    it('should function while eth_coinbase', () => {
-      const eth = new Eth(provider);
+    it('should function while vap_coinbase', () => {
+      const vap = new Vap(provider);
 
-      eth.coinbase((err, result) => {
+      vap.coinbase((err, result) => {
         assert.equal(err, null);
         assert.equal(typeof result, 'string');
         assert.equal(util.getBinarySize(result), 42);
@@ -40,9 +40,9 @@ describe('ethjs-query', () => {
     });
 
     it('should function while get_accounts', (done) => {
-      const eth = new Eth(provider);
+      const vap = new Vap(provider);
 
-      eth.accounts((err, result) => {
+      vap.accounts((err, result) => {
         assert.equal(err, null);
         assert.equal(typeof result, 'object');
         assert.equal(Array.isArray(result), true);
@@ -53,10 +53,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_blockNumber', (done) => {
-      const eth = new Eth(provider);
+    it('should function while vap_blockNumber', (done) => {
+      const vap = new Vap(provider);
 
-      eth.blockNumber((err, result) => {
+      vap.blockNumber((err, result) => {
         assert.equal(err, null);
         assert.equal(typeof result, 'object');
         assert.equal(result.toNumber() >= 0, true);
@@ -64,8 +64,8 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_compileSolidity', (done) => {
-      const eth = new Eth(provider);
+    it('should function while vap_compileSolidity', (done) => {
+      const vap = new Vap(provider);
       const testSolidity = `pragma solidity ^0.4.0;
 
       /// @title Voting with delegation.
@@ -77,7 +77,7 @@ describe('ethjs-query', () => {
       }
       `;
 
-      eth.compileSolidity(testSolidity, (err, result) => {
+      vap.compileSolidity(testSolidity, (err, result) => {
         assert.equal(err, null);
         assert.equal(typeof result, 'object');
         assert.equal(typeof result.code, 'string');
@@ -87,9 +87,9 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_estimateGas', (done) => {
-      const eth = new Eth(provider);
-      eth.accounts((accountsError, accounts) => {
+    it('should function while vap_estimateGas', (done) => {
+      const vap = new Vap(provider);
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -101,7 +101,7 @@ describe('ethjs-query', () => {
           data: '0x',
         };
 
-        eth.estimateGas(testTransactionObject, (err, result) => {
+        vap.estimateGas(testTransactionObject, (err, result) => {
           assert.equal(err, null);
           assert.equal(typeof result, 'object');
           assert.equal(typeof result.toString(10), 'string');
@@ -111,10 +111,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_gasPrice', (done) => {
-      const eth = new Eth(provider);
+    it('should function while vap_gasPrice', (done) => {
+      const vap = new Vap(provider);
 
-      eth.gasPrice((err, result) => {
+      vap.gasPrice((err, result) => {
         assert.equal(err, null);
         assert.equal(typeof result, 'object');
         assert.equal(result.toNumber() > 0, true);
@@ -122,19 +122,19 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getBalance', (done) => {
-      const eth = new Eth(provider);
+    it('should function while vap_getBalance', (done) => {
+      const vap = new Vap(provider);
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
-        eth.getBalance(accounts[0], (err, result) => {
+        vap.getBalance(accounts[0], (err, result) => {
           assert.equal(err, null);
           assert.equal(typeof result, 'object');
           assert.equal(result.toNumber() > 0, true);
 
-          eth.getBalance(accounts[0], 'latest', (err2, result2) => {
+          vap.getBalance(accounts[0], 'latest', (err2, result2) => {
             assert.equal(err2, null);
             assert.equal(typeof result2, 'object');
             assert.equal(result2.toNumber() > 0, true);
@@ -144,10 +144,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getBlockByNumber', (done) => { // eslint-disable-line
-      const eth = new Eth(provider);
+    it('should function while vap_getBlockByNumber', (done) => { // eslint-disable-line
+      const vap = new Vap(provider);
 
-      eth.getBlockByNumber(0, true, (blockError, result) => {
+      vap.getBlockByNumber(0, true, (blockError, result) => {
         assert.equal(blockError, null);
         assert.equal(typeof result, 'object');
         assert.equal(util.getBinarySize(result.hash), 66);
@@ -160,14 +160,14 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getBlockByHash', (done) => {
-      const eth = new Eth(provider);
+    it('should function while vap_getBlockByHash', (done) => {
+      const vap = new Vap(provider);
 
-      eth.getBlockByNumber(0, true, (blockError, block) => {
+      vap.getBlockByNumber(0, true, (blockError, block) => {
         assert.equal(blockError, null);
         assert.equal(typeof block, 'object');
 
-        eth.getBlockByHash(block.hash, true, (error, result) => {
+        vap.getBlockByHash(block.hash, true, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'object');
           assert.equal(util.getBinarySize(result.hash), 66);
@@ -181,16 +181,16 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getCode', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_getCode', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
       done();
     });
 
-    it('should function while eth_getCompilers', (done) => {
+    it('should function while vap_getCompilers', (done) => {
       /*
-      const eth = new Eth(provider); // eslint-disable-line
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.getCompilers((compilerError, compilerResilt) => {
+      vap.getCompilers((compilerError, compilerResilt) => {
         console.log(compilerError, compilerResilt);
 
         // assert.equal(error, null);
@@ -203,10 +203,10 @@ describe('ethjs-query', () => {
       done();
     });
 
-    it('should function while eth_hashrate', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_hashrate', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.hashrate((error, result) => {
+      vap.hashrate((error, result) => {
         assert.equal(error, null);
         assert.equal(typeof result, 'object');
         assert.equal(result.toNumber(10) >= 0, true);
@@ -215,10 +215,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_mining', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_mining', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.mining((error, result) => {
+      vap.mining((error, result) => {
         assert.equal(error, null);
         assert.equal(typeof result, 'boolean');
 
@@ -226,14 +226,14 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getTransactionCount', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_getTransactionCount', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
-        eth.getTransactionCount(accounts[0], (error, result) => {
+        vap.getTransactionCount(accounts[0], (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'object');
           assert.equal(result.toNumber(10) >= 0, true);
@@ -243,10 +243,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getTransactionByBlockHashAndIndex', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_getTransactionByBlockHashAndIndex', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -257,16 +257,16 @@ describe('ethjs-query', () => {
           data: '0x',
         };
 
-        eth.sendTransaction(testTransaction, (error, result) => {
+        vap.sendTransaction(testTransaction, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result), 66);
 
-          eth.getTransactionReceipt(result, (receiptError, receipt) => {
+          vap.getTransactionReceipt(result, (receiptError, receipt) => {
             assert.equal(receiptError, null);
             assert.equal(typeof receipt, 'object');
 
-            eth.getTransactionByBlockHashAndIndex(receipt.blockHash, 0, (blockError, block) => {
+            vap.getTransactionByBlockHashAndIndex(receipt.blockHash, 0, (blockError, block) => {
               assert.equal(blockError, null);
               assert.equal(typeof block, 'object');
               assert.equal(util.getBinarySize(block.blockHash), 66);
@@ -282,10 +282,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getTransactionByBlockNumberAndIndex', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_getTransactionByBlockNumberAndIndex', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -296,16 +296,16 @@ describe('ethjs-query', () => {
           data: '0x',
         };
 
-        eth.sendTransaction(testTransaction, (error, result) => {
+        vap.sendTransaction(testTransaction, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result), 66);
 
-          eth.getTransactionReceipt(result, (receiptError, receipt) => {
+          vap.getTransactionReceipt(result, (receiptError, receipt) => {
             assert.equal(receiptError, null);
             assert.equal(typeof receipt, 'object');
 
-            eth.getTransactionByBlockNumberAndIndex(2, 0, (blockError, block) => {
+            vap.getTransactionByBlockNumberAndIndex(2, 0, (blockError, block) => {
               assert.equal(blockError, null);
               assert.equal(typeof block, 'object');
               assert.equal(util.getBinarySize(block.blockHash), 66);
@@ -321,10 +321,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_sendTransaction', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_sendTransaction', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -335,7 +335,7 @@ describe('ethjs-query', () => {
           data: '0x',
         };
 
-        eth.sendTransaction(testTransaction, (error, result) => {
+        vap.sendTransaction(testTransaction, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result), 66);
@@ -345,10 +345,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_sendTransaction with contract', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_sendTransaction with contract', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -358,7 +358,7 @@ describe('ethjs-query', () => {
           data: '606060405234610000575b61016a806100186000396000f360606040526000357c010000000000000000000000000000000000000000000000000000000090048063119c56bd1461004e57806360fe47b11461008e5780636d4ce63c146100c1575b610000565b346100005761005b6100e4565b604051808381526020018273ffffffffffffffffffffffffffffffffffffffff1681526020019250505060405180910390f35b34610000576100a960048080359060200190919050506100f5565b60405180821515815260200191505060405180910390f35b34610000576100ce61015f565b6040518082815260200191505060405180910390f35b60006000610d7d91503390505b9091565b6000816000819055507f10e8e9bc5a1bde3dd6bb7245b52503fcb9d9b1d7c7b26743f82c51cc7cce917d60005433604051808381526020018273ffffffffffffffffffffffffffffffffffffffff1681526020019250505060405180910390a1600190505b919050565b600060005490505b9056',
         };
 
-        eth.sendTransaction(testTransaction, (error, result) => {
+        vap.sendTransaction(testTransaction, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result), 66);
@@ -368,16 +368,16 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_sign', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_sign', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
         const testTxData = '0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470';
 
-        eth.sign(accounts[0], testTxData, (error, result) => {
+        vap.sign(accounts[0], testTxData, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result) > 0, true);
@@ -387,10 +387,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while eth_getTransactionReceipt', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while vap_getTransactionReceipt', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -401,13 +401,13 @@ describe('ethjs-query', () => {
           data: '0x',
         };
 
-        eth.sendTransaction(testTransaction, (error, result) => {
+        vap.sendTransaction(testTransaction, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result), 66);
 
           setTimeout(() => {
-            eth.getTransactionReceipt(result, (receiptError, receipt) => {
+            vap.getTransactionReceipt(result, (receiptError, receipt) => {
               assert.equal(receiptError, null);
               assert.equal(typeof receipt, 'object');
 
@@ -425,10 +425,10 @@ describe('ethjs-query', () => {
       });
     });
 
-    it('should function while deploy, use contract via eth_call, eth_getCode', (done) => {
-      const eth = new Eth(provider); // eslint-disable-line
+    it('should function while deploy, use contract via vap_call, vap_getCode', (done) => {
+      const vap = new Vap(provider); // eslint-disable-line
 
-      eth.accounts((accountsError, accounts) => {
+      vap.accounts((accountsError, accounts) => {
         assert.equal(accountsError, null);
         assert.equal(typeof accounts, 'object');
 
@@ -440,13 +440,13 @@ describe('ethjs-query', () => {
 
         const contractABI = [{'constant': false,'inputs': [],'name': 'setcompeltereturn','outputs': [{'name': '_newValue','type': 'uint256'},{'name': '_sender','type': 'address'}],'payable': false,'type': 'function'},{'constant': false,'inputs': [{'name': '_value','type': 'uint256'}],'name': 'set','outputs': [{'name': '','type': 'bool'}],'payable': false,'type': 'function'},{'constant': false,'inputs': [],'name': 'get','outputs': [{'name': 'storeValue','type': 'uint256'}],'payable': false,'type': 'function'},{'anonymous':false,'inputs':[{'indexed':false,'name':'_newValue','type':'uint256'},{'indexed':false,'name':'_sender','type':'address'}],'name':'SetComplete','type':'event'}]; // eslint-disable-line
 
-        eth.sendTransaction(testContractTransaction, (error, result) => {
+        vap.sendTransaction(testContractTransaction, (error, result) => {
           assert.equal(error, null);
           assert.equal(typeof result, 'string');
           assert.equal(util.getBinarySize(result), 66);
 
           setTimeout(() => {
-            eth.getTransactionReceipt(result, (receiptError, receipt) => {
+            vap.getTransactionReceipt(result, (receiptError, receipt) => {
               assert.equal(receiptError, null);
               assert.equal(typeof receipt, 'object');
 
@@ -466,7 +466,7 @@ describe('ethjs-query', () => {
                 data: abi.encodeMethod(contractABI[1], [uintValue]),
               };
 
-              eth.sendTransaction(setMethodTransaction, (setMethodError, setMethodTx) => {
+              vap.sendTransaction(setMethodTransaction, (setMethodError, setMethodTx) => {
                 assert.equal(setMethodError, null);
                 assert.equal(typeof setMethodTx, 'string');
                 assert.equal(util.getBinarySize(setMethodTx), 66);
@@ -477,13 +477,13 @@ describe('ethjs-query', () => {
                     data: abi.encodeMethod(contractABI[2], []),
                   };
 
-                  eth.call(callMethodTransaction, (callError, callResult) => { // eslint-disable-line
+                  vap.call(callMethodTransaction, (callError, callResult) => { // eslint-disable-line
                     assert.equal(setMethodError, null);
                     const decodedUint = abi.decodeMethod(contractABI[2], callResult);
 
                     assert.equal(decodedUint[0].toNumber(10), uintValue);
 
-                    eth.getCode(receipt.contractAddress, 'latest', (codeError, codeResult) => {
+                    vap.getCode(receipt.contractAddress, 'latest', (codeError, codeResult) => {
                       assert.equal(codeError, null);
                       assert.equal(typeof codeResult, 'string');