+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-# 0.0.2 -- handle eth_getFilterChanges during Block and Pending Tx filter
+# 0.0.2 -- handle vap_getFilterChanges during Block and Pending Tx filter
 
 1. handle getFilterChanges during BlockFilter and PendingTxFilter.
 
 
-# 0.0.1 -- ethjs-query
+# 0.0.1 -- vapjs-query
 
 1. Basic testing
 2. Basic docs